Summary request folder comes Output folder and namespaces change. Notifications refactoring.
