+++ app/Http/Controllers/Web/SpaceCalculator/OutputsController.php
@@ -8,13 +8,13 @@ use App\Actions\MagicLinks\SendAction;
 use App\Actions\Users\CreateAction;
 use App\Actions\Users\UpdateProfileAction;
 use App\Http\Controllers\WebController;
-use App\Http\Requests\Web\SpaceCalculator\Summary\PostProfileRequest;
-use App\Http\Requests\Web\SpaceCalculator\Summary\PostSummaryRequest;
+use App\Http\Requests\Web\SpaceCalculator\Outputs\PostProfileRequest;
+use App\Http\Requests\Web\SpaceCalculator\Outputs\PostSummaryRequest;
 use App\Jobs\Enquiries\TransmitToHubSpotJob;
 use App\Models\SpaceCalculatorInput;
 use App\Models\User;
-use App\Notifications\DetailedNotification;
-use App\Notifications\SummaryNotification;
+use App\Notifications\SpaceCalculator\DetailedNotification;
+use App\Notifications\SpaceCalculator\SummaryNotification;
 use App\Services\SpaceCalculator\Calculator;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
